pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "py-rustycluster-client"
version = "1.0.0"
description = "Enterprise-grade Python gRPC client for the RustyCluster distributed key-value store"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [{ name = "RustyCluster", email = "support@rustycluster.io" }]
keywords = ["grpc", "key-value", "distributed", "redis", "rustycluster"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Database",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]

dependencies = [
    "grpcio>=1.60.0",
    "protobuf>=4.25.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
async = [
    "grpcio>=1.60.0",  # grpc.aio is included in grpcio
]
dev = [
    "grpcio-tools>=1.60.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.7.0",
    "ruff>=0.1.0",
    "types-protobuf>=4.24.0",
]

[project.urls]
Homepage = "https://github.com/rustycluster/rustycluster-python-client"
Documentation = "https://docs.rustycluster.io/python-client"
Repository = "https://github.com/rustycluster/rustycluster-python-client"
Issues = "https://github.com/rustycluster/rustycluster-python-client/issues"

[tool.setuptools.packages.find]
include = ["rustycluster*"]

[tool.setuptools.package-data]
rustycluster = ["py.typed"]

# ──────────────────────────────────────────────
# Ruff linting
# ──────────────────────────────────────────────
[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "ANN", "C4", "SIM"]
ignore = [
    "ANN101",  # Missing type annotation for self
    "ANN102",  # Missing type annotation for cls
    "ANN401",  # Dynamically typed expressions (Any)
    "B008",    # Do not perform function calls in default arguments
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["ANN", "S101"]

# ──────────────────────────────────────────────
# Mypy type checking
# ──────────────────────────────────────────────
[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
exclude = ["rustycluster/proto/"]

# ──────────────────────────────────────────────
# Pytest
# ──────────────────────────────────────────────
[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["rustycluster"]
omit = ["rustycluster/proto/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
    "@abstractmethod",
]
